+++ carp_audio_package/lib/audio.g.dart
@@ -83,8 +83,8 @@ AudioMeasure _$AudioMeasureFromJson(Map<String, dynamic> json) {
     json['type'] == null
         ? null
         : MeasureType.fromJson(json['type'] as Map<String, dynamic>),
-    name: json['name'],
-    enabled: json['enabled'],
+    name: json['name'] as String,
+    enabled: json['enabled'] as bool,
     studyId: json['study_id'] as String,
   )
     ..c__ = json['c__'] as String
@@ -116,10 +116,14 @@ NoiseMeasure _$NoiseMeasureFromJson(Map<String, dynamic> json) {
     json['type'] == null
         ? null
         : MeasureType.fromJson(json['type'] as Map<String, dynamic>),
-    name: json['name'],
-    enabled: json['enabled'],
-    frequency: json['frequency'],
-    duration: json['duration'],
+    name: json['name'] as String,
+    enabled: json['enabled'] as bool,
+    frequency: json['frequency'] == null
+        ? null
+        : Duration(microseconds: json['frequency'] as int),
+    duration: json['duration'] == null
+        ? null
+        : Duration(microseconds: json['duration'] as int),
     samplingRate: json['sampling_rate'] as int,
   )
     ..c__ = json['c__'] as String
@@ -142,8 +146,8 @@ Map<String, dynamic> _$NoiseMeasureToJson(NoiseMeasure instance) {
   writeNotNull('name', instance.name);
   writeNotNull('enabled', instance.enabled);
   writeNotNull('configuration', instance.configuration);
-  writeNotNull('frequency', instance.frequency);
-  writeNotNull('duration', instance.duration);
+  writeNotNull('frequency', instance.frequency?.inMicroseconds);
+  writeNotNull('duration', instance.duration?.inMicroseconds);
   writeNotNull('sampling_rate', instance.samplingRate);
   return val;
 }