+++ carp_connectivity_package/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.8.0
+* upgrade to `carp_mobile_sensing` v. 0.8.x
+
 ## 0.6.6
 * upgrade to `carp_mobile_sensing` v. 0.6.6